third_party/blink/web_tests/editing/selection/modify_extend/extend_backward_line_to_smaller_line.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>

<script>
  test(
    () =>
      assert_selection(
        [
          '<div contenteditable style="width: 3ch; white-space: pre-wrap;">',
          '<bdo style="font-family: monospace; width: 3ch; white-space: pre-wrap;">',
          " Abc BC DEf|</bdo></div>",
        ].join(" "),
        (selection) => {
          selection.modify("extend", "backward", "line");
          selection.modify("extend", "backward", "line");
        },
        [
          '<div contenteditable style="width: 3ch; white-space: pre-wrap;">',
          '<bdo style="font-family: monospace; width: 3ch; white-space: pre-wrap;">',
          " Abc| BC DEf^</bdo></div>",
        ].join(" ")
      ),
    "Extend selection backward twice by line in a white space wrapped paragraph."
  );
</script>
